Update collection.json oceanic fronts and currents
This product - the oceanic fronts, is about 3 areas in the Atlantic. I reserved this one for the Gibraltar and I am going to create two collections for the other two ones. I updated the bounding box, time period and description of the product, as well as the hyperlink to match directly to the gibraltar product.