WOC_Update_v2_23122024
Tried to resolve most of the comments.
Created Sargassum variable, with description and products, updated fronts and oil spill.
One that still remains is about the google drive/documentation of the products. It can be found on their own page, which is linked to the product, but of course I can change it.